# fix(internal/spec): do a pointer comparison for table objects instead of a deep compare