The conversion from an AST to a spec would evaluate the AST and then
return a spec from the side effects that were related to table objects.

In order to do this, the code that converted a table object into a spec
would also check to see if that table object had already been seen
before.

This caused a performance problem because the `Equal` method for a
`TableObject` would recursively check the `TableObject` which also
included checking the parents. If you had a reasonably complex query,
this could take a very long time as that code was not very efficient and
involved lots of small memory allocations.

This changes it so that instead of recursively checking, it just checks
the memory location which should be the same when they are duplicates
anyway.

### Done checklist
- [x] docs/SPEC.md updated
- [x] Test cases written