@jsternberg 2019-09-11
Benchmarks:

```
benchmark              old ns/op      new ns/op     delta
Benchmark_Build-12     7707517459     34494225      -99.55%

benchmark              old allocs     new allocs     delta
Benchmark_Build-12     140373067      391713         -99.72%

benchmark              old bytes      new bytes     delta
Benchmark_Build-12     4959824792     22660985      -99.54%
```

I manually verified with code coverage that the benchmark covered that piece of code and took the true conditional path.